fix(suite): move types to components and delete unused ones
