+++ packages/components/src/components/form/inputTypes.ts
@@ -1,0 +1,3 @@
+export type InputState = 'success' | 'warning' | 'error';
+
+export type InputSize = 'small' | 'large';
