+++ packages/suite/src/components/suite/DropZone.tsx
@@ -1,9 +1,8 @@
 import { useRef, useCallback, useMemo, useState, MouseEvent, DragEvent, ChangeEvent } from 'react';
 import styled from 'styled-components';
-import { P, Icon } from '@trezor/components';
+import { P, Icon, IconType } from '@trezor/components';
 import { Translation } from 'src/components/suite';
 import type { ExtendedMessageDescriptor } from 'src/types/suite';
-import { IconType } from '@trezor/components/src/support/types';
 
 interface DropZoneProps {
     // 'accept' attribute for underlying HTML file input
