+++ packages/suite/styled.d.ts
@@ -1,8 +1,8 @@
 // import original module declarations
 import 'styled-components';
-import { types } from '@trezor/components';
+import { SuiteThemeColors } from '@trezor/components';
 import { Colors } from '@trezor/theme';
 
 declare module 'styled-components' {
-    export interface DefaultTheme extends types.SuiteThemeColors, Colors {}
+    export interface DefaultTheme extends SuiteThemeColors, Colors {}
 }
